Finish sizer interface (#3752)
* Use int64_t for CalculateSizeBytes

This avoids a bunch of nasty casting/checking in other parts of the code

* Use the Sizer for MutationBatches

* Don't double-count document key size.

+++ Firestore/core/src/firebase/firestore/local/memory_remote_document_cache.h
@@ -58,7 +58,7 @@ class MemoryRemoteDocumentCache : public RemoteDocumentCache {
       FSTMemoryLRUReferenceDelegate* reference_delegate,
       model::ListenSequenceNumber upper_bound);
 
-  size_t CalculateByteSize(const Sizer& sizer);
+  int64_t CalculateByteSize(const Sizer& sizer);
 
  private:
   /** Underlying cache of documents. */
